pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "autocalver"]

[project]
name = "autocalver"
dynamic = ["version"]
dependencies = ["tomli", "python-dateutil"]
readme = "README.md"

[project.entry-points."setuptools.finalize_distribution_options"]
autocalver = "autocalver.integration:guess_version"

[project.entry-points."setuptools.file_finders"]
autocalver = "autocalver.integration:find_files"

[project.optional-dependencies]
tests = ["coverage", "pytest"]
lint = ["black", "mypy", "flake8", "pylint"]
mypy = ["mypy", "types-python-dateutil"]
docs = ["sphinx"]


[tool.autocalver]
use = true
log = "git-log-head"
log_command = "git log -n 1 --date=iso"
is_main_var = "GITHUB_REF"
is_main_match = ".*/trunk$"
